+++ test/acceptance/ported/raster.js
@@ -41,7 +41,7 @@ describe('raster', function() {
           assert.response(server, {
               url: '/database/windshaft_test/layergroup',
               method: 'POST',
-              headers: {'Content-Type': 'application/json' },
+              headers: { host: 'localhost', 'Content-Type': 'application/json' },
               data: JSON.stringify(mapconfig)
           }, {}, function(res, err) { next(err, res); });
         },
@@ -66,7 +66,8 @@ describe('raster', function() {
           assert.response(server, {
               url: '/database/windshaft_test/layergroup/' + expected_token + '/0/0/0.png',
               method: 'GET',
-              encoding: 'binary'
+              encoding: 'binary',
+              headers: { host: 'localhost' }
           }, {}, function(res, err) { next(err, res); });
         },
         function check_response(err, res) {
@@ -124,6 +125,7 @@ describe('raster', function() {
                 url: '/database/windshaft_test/layergroup',
                 method: 'POST',
                 headers: {
+                    host: 'localhost',
                     'Content-Type': 'application/json'
                 },
                 data: JSON.stringify(mapconfig)