+++ test/acceptance/ported/support/ported_server_options.js
@@ -1,7 +1,6 @@
 var _ = require('underscore');
 var serverOptions = require('../../../../lib/cartodb/server_options');
 var mapnik = require('windshaft').mapnik;
-var LayergroupToken = require('../../../../lib/cartodb/models/layergroup-token');
 var OverviewsQueryRewriter = require('../../../../lib/cartodb/utils/overviews_query_rewriter');
 var overviewsQueryRewriter = new OverviewsQueryRewriter({
   zoom_level: 'CDB_ZoomFromScale(!scale_denominator!)'
@@ -47,48 +46,5 @@ module.exports = _.extend({}, serverOptions, {
     enable_cors: global.environment.enable_cors,
     unbuffered_logging: true, // for smoother teardown from tests
     log_format: null, // do not log anything
-    afterLayergroupCreateCalls: 0,
-    useProfiler: true,
-    req2params: function(req, res, callback){
-
-        if ( req.query.testUnexpectedError ) {
-            return callback('test unexpected error');
-        }
-
-        // this is in case you want to test sql parameters eg ...png?sql=select * from my_table limit 10
-        req.params =  _.extend({}, req.params);
-
-        if (req.params.token) {
-            req.params.token = LayergroupToken.parse(req.params.token).token;
-        }
-
-        _.extend(req.params, req.query);
-        req.params.user = 'localhost';
-        res.locals.user = 'localhost';
-
-        req.params.dbhost = global.environment.postgres.host;
-        req.params.dbport = req.params.dbport || global.environment.postgres.port;
-
-        req.params.dbuser = 'test_windshaft_publicuser';
-        if (req.params.dbname !== 'windshaft_test2') {
-            req.params.dbuser = 'test_windshaft_cartodb_user_1';
-        }
-        req.params.dbname = 'test_windshaft_cartodb_user_1_db';
-
-        // add all params to res.locals
-        res.locals = _.extend({}, req.params);
-
-
-        // increment number of calls counter
-        global.req2params_calls = global.req2params_calls ? global.req2params_calls + 1 : 1;
-
-        // send the finished req object on
-        callback(null,req);
-    },
-    afterLayergroupCreate: function(req, cfg, res, callback) {
-        res.layercount = cfg.layers.length;
-//        config.afterLayergroupCreateCalls++;
-        callback(null);
-    }
-
+    useProfiler: true
 });